requirements.txt
# =============================================================================
# HYPERVIEW2 Challenge - Python Dependencies
# =============================================================================

# Core Scientific Computing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Explainable AI
shap>=0.42.0

# Configuration
pyyaml>=6.0

# Model Persistence
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars (optional but recommended)
tqdm>=4.65.0

# Jupyter Support (optional - for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0